solution-checker/requirements.txt
pymongo==3.12.0
python-decouple==3.4

